Move system_struct_to_node_substates to system_type_checker
