+++ transaction-scenarios/generated-examples/fungible_resource/002--fungible-max-div-mint.rtm
@@ -4,7 +4,7 @@ CALL_METHOD
     Decimal("5000")
 ;
 MINT_FUNGIBLE
-    Address("resource_sim1tk6xaxxcxdnhfusk0w8sw5f39vn8utt5sl53eyj44rhzd97m4rktlu")
+    Address("resource_sim1t5y7vqwyeef2sqrja6m5k6ccne4vz09pazfgppdmjj3urwkqr2794q")
     Decimal("100")
 ;
 CALL_METHOD
