+++ transaction-scenarios/generated-examples/fungible_resource/003--fungible-max-div-burn.rtm
@@ -6,11 +6,11 @@ CALL_METHOD
 CALL_METHOD
     Address("account_sim16996e320lnez82q6430eunaz9l3n5fnwk6eh9avrmtmj22e7jmhemw")
     "withdraw"
-    Address("resource_sim1tk6xaxxcxdnhfusk0w8sw5f39vn8utt5sl53eyj44rhzd97m4rktlu")
+    Address("resource_sim1t5y7vqwyeef2sqrja6m5k6ccne4vz09pazfgppdmjj3urwkqr2794q")
     Decimal("10")
 ;
 TAKE_ALL_FROM_WORKTOP
-    Address("resource_sim1tk6xaxxcxdnhfusk0w8sw5f39vn8utt5sl53eyj44rhzd97m4rktlu")
+    Address("resource_sim1t5y7vqwyeef2sqrja6m5k6ccne4vz09pazfgppdmjj3urwkqr2794q")
     Bucket("to_burn")
 ;
 BURN_RESOURCE
