+++ transaction-scenarios/generated-examples/fungible_resource/018--fungible-min-div-transfer-wrong-granularity.rtm
@@ -6,7 +6,7 @@ CALL_METHOD
 CALL_METHOD
     Address("account_sim16996e320lnez82q6430eunaz9l3n5fnwk6eh9avrmtmj22e7jmhemw")
     "withdraw"
-    Address("resource_sim1thruzvhydk0p2dgsjmxfuvre7dtsxap8zjkfm52znuwl6277qr4qlp")
+    Address("resource_sim1t447gzsnsflhndp6gy2jqm5mdyffqncx9pqegxu03w3as765g7x8wl")
     Decimal("0.0001")
 ;
 CALL_METHOD
