+++ transaction-scenarios/generated-examples/metadata/001--metadata-create-package-with-metadata.rtm
@@ -471,7 +471,7 @@ PUBLISH_PACKAGE_ADVANCED
             )
         )
     )
-    Blob("f824ffbd7cdb8d442830c4799f6c1f76707170e1d8294c026a6e205888fbf495")
+    Blob("37ec0b71971f6e6ce9e1adb28d67214be5b4a5515344b7ced08ac4e359741ab4")
     Map<String, Tuple>(
         "" => Tuple(
             Enum<1u8>(
