+++ transaction-scenarios/generated-examples/metadata/002--metadata-create-component-with-metadata.rtm
@@ -4,7 +4,7 @@ CALL_METHOD
     Decimal("5000")
 ;
 ALLOCATE_GLOBAL_ADDRESS
-    Address("package_sim1p4vp4w3dnjz60fn69kzcn0qgn6tcgg8839lh3py825mwz6hhenm99u")
+    Address("package_sim1p46sga6yesuvqn5memflf6zj3wnqescka4tse00staqk3cddm0mh4h")
     "MetadataTest"
     AddressReservation("metadata_component_address_reservation")
     NamedAddress("metadata_component_address")
@@ -14,7 +14,7 @@ CALL_METHOD
     "free"
 ;
 CALL_FUNCTION
-    Address("package_sim1p4vp4w3dnjz60fn69kzcn0qgn6tcgg8839lh3py825mwz6hhenm99u")
+    Address("package_sim1p46sga6yesuvqn5memflf6zj3wnqescka4tse00staqk3cddm0mh4h")
     "MetadataTest"
     "new_with_address"
     AddressReservation("metadata_component_address_reservation")
