+++ transaction-scenarios/generated-examples/metadata/006--metadata-update-updatable-metadata-succeeds.rtm
@@ -4,7 +4,7 @@ CALL_METHOD
     Decimal("5000")
 ;
 SET_METADATA
-    Address("resource_sim1t4xmn2lk7cqgrahyq8qly378w8sjccfcd92f2amqw26ks39gy9n3d8")
+    Address("resource_sim1thv9xxqllmdjygwpgur8lurn6f9kx6hlmdq6e6eaw883jsryaysycg")
     "locked_later"
     Enum<1u8>(
         true
