+++ transaction-scenarios/generated-examples/metadata/007--metadata-lock-metadata.rtm
@@ -4,6 +4,6 @@ CALL_METHOD
     Decimal("5000")
 ;
 LOCK_METADATA
-    Address("resource_sim1t4xmn2lk7cqgrahyq8qly378w8sjccfcd92f2amqw26ks39gy9n3d8")
+    Address("resource_sim1thv9xxqllmdjygwpgur8lurn6f9kx6hlmdq6e6eaw883jsryaysycg")
     "locked_later"
 ;
