+++ transaction-scenarios/generated-examples/non_fungible_resource/010--non-fungible-resource-recall-frozen-vault.rtm
@@ -4,7 +4,7 @@ CALL_METHOD
     Decimal("5000")
 ;
 RECALL_NON_FUNGIBLES_FROM_VAULT
-    Address("internal_vault_sim1np3haxqlc8xwx08j9wt0k8nkzgmq43csg2xwqg686c76lgg5dst70f")
+    Address("internal_vault_sim1nz6g6eyn7e3etkggcp2jwqkdq3nf4mzdhdgunaje7sphpmex6gfjha")
     Array<NonFungibleLocalId>(
         NonFungibleLocalId("#120#")
     )
