+++ transaction-scenarios/generated-examples/non_fungible_resource/017--non-fungible-transfer-metadata-standard-nfs.rtm
@@ -6,7 +6,7 @@ CALL_METHOD
 CALL_METHOD
     Address("account_sim16996e320lnez82q6430eunaz9l3n5fnwk6eh9avrmtmj22e7jmhemw")
     "withdraw_non_fungibles"
-    Address("resource_sim1ntr20r9sdw9kd9xfzqf5af7n5d6euz36gfyn6udrjqxkznrm9yc2hz")
+    Address("resource_sim1ntvvup84xzj0zjc92g07h2axttd0cxu3waytwhahsmpvwj2fsanq45")
     Array<NonFungibleLocalId>(
         NonFungibleLocalId("#4#"),
         NonFungibleLocalId("#8#")
