+++ transaction-scenarios/generated-examples/non_fungible_resource/018--non-fungible-create-resource-with-supply-with-complex-data.rtm
@@ -194,7 +194,7 @@ CREATE_NON_FUNGIBLE_RESOURCE_WITH_INITIAL_SUPPLY
         NonFungibleLocalId("#1#") => Tuple(
             Tuple(
                 100u64,
-                NonFungibleGlobalId("resource_sim1ntr20r9sdw9kd9xfzqf5af7n5d6euz36gfyn6udrjqxkznrm9yc2hz:#8#"),
+                NonFungibleGlobalId("resource_sim1ntvvup84xzj0zjc92g07h2axttd0cxu3waytwhahsmpvwj2fsanq45:#8#"),
                 "Some string which could be made long for test cases",
                 Tuple(
                     42u8,
