+++ transaction-scenarios/generated-examples/non_fungible_resource/019--non-fungible-mutate-data.rtm
@@ -4,14 +4,14 @@ CALL_METHOD
     Decimal("5000")
 ;
 CALL_METHOD
-    Address("resource_sim1ntrzuv52mcnukjhje87kaeakp3r08dh4l84ldk2r9rvc67j8drtpux")
+    Address("resource_sim1n297yf5k0d423xppgvj9wc4qucm409lmy9rmmnpzzfurqxkk7s6ql5")
     "update_non_fungible_data"
     NonFungibleLocalId("#1#")
     "mutable_long_name_for_data_to_try_and_stretch_the_bounds_of_what_is_possible_in_user_interfaces"
     "This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! This is a long string with repeats of length 50!! "
 ;
 CALL_METHOD
-    Address("resource_sim1ntrzuv52mcnukjhje87kaeakp3r08dh4l84ldk2r9rvc67j8drtpux")
+    Address("resource_sim1n297yf5k0d423xppgvj9wc4qucm409lmy9rmmnpzzfurqxkk7s6ql5")
     "update_non_fungible_data"
     NonFungibleLocalId("#1#")
     "mutable_inner_enum"
