+++ transaction-scenarios/generated-examples/radiswap/007--radiswap-remove-tokens.rtm
@@ -6,15 +6,15 @@ CALL_METHOD
 CALL_METHOD
     Address("account_sim168j3paqgngj74yzaljq4n422rtsmupaec3wnqq5425fd85cnd8xmdz")
     "withdraw"
-    Address("resource_sim1th6yu877es2lr66269m8n58x2d52685pr55gmsrr3rqyxv9lxetyu5")
+    Address("resource_sim1tk3dlzqewq42r7y70686ge6zn978yjc2ntutyztcwflf3l7ew8mdan")
     Decimal("100")
 ;
 TAKE_ALL_FROM_WORKTOP
-    Address("resource_sim1th6yu877es2lr66269m8n58x2d52685pr55gmsrr3rqyxv9lxetyu5")
+    Address("resource_sim1tk3dlzqewq42r7y70686ge6zn978yjc2ntutyztcwflf3l7ew8mdan")
     Bucket("pool_units")
 ;
 CALL_METHOD
-    Address("component_sim1cz0q8qdutuw4smq3h4aq3mes5w6we28k24umk2ym47uekgsvs9d2v6")
+    Address("component_sim1cq49w9xq09scf526j6efeq4cgqw6pnnc3pexu4kwvfgckpj7hcrpv4")
     "remove_liquidity"
     Bucket("pool_units")
 ;
